Adding google closure stylesheet repo as local file
This is not ideal. In framework closure-stylesheets repo is in a local file, not the central repo. For some reason when note uses newer versions of Aura framework it looks for and can't find the stylesheets repo so moving it to the same location in note project as it is in the framework. We should find a better way to do this to avoid manually copying over files from the framework to note whenever the stylesheets repo is updated.